# apps/server/openapi.v2.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Ashyq Bilim API",
    "description": "University LMS / MOOC platform API (v2, Rust rewrite).",
    "license": {
      "name": ""
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v2/collections": {
      "get": {
        "tags": [
          "collections"
        ],
        "summary": "Newest-first collection listing: public plus the caller's own.",
        "operationId": "list_collections",
        "parameters": [
          {
            "name": "cursor",
            "in": "query",
            "description": "next_cursor from the previous page",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/CollectionId"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Page size, 1..=100 (default 20)",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Page of collections",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CollectionPage"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "collections"
        ],
        "summary": "Create a collection (requires `collection:create:platform`); every\nattached course must be readable by the caller.",
        "operationId": "create_collection",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCollectionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Collection"
                }
              }
            }
          },
          "403": {
            "description": "Missing permission",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
    "/api/v2/collections/{id}": {
      "get": {
        "tags": [
          "collections"
        ],
        "summary": "One collection with its member courses (404 when invisible).",
        "operationId": "get_collection",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Collection id",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/CollectionId"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Collection",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Collection"
                }
              }
            }
          },
          "404": {
            "description": "Unknown or inaccessible",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "collections"
        ],
        "summary": "Delete a collection (membership rows cascade; courses stay).",
        "operationId": "delete_collection",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Collection id",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/CollectionId"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Deleted"
          },
          "403": {
            "description": "No delete access",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "collections"
        ],
        "summary": "Partial update; `courses` replaces the whole membership when present.",
        "operationId": "update_collection",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Collection id",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/CollectionId"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCollectionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Collection"
                }
              }
            }
          },
          "403": {
            "description": "No write access",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Collection": {
        "type": "object",
        "required": [
          "id",
          "name",
          "description",
          "public",
          "courses",
          "created_at_unix",
          "updated_at_unix"
        ],
        "properties": {
          "courses": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Course"
            },
            "description": "Member courses visible to the caller, in collection order."
          },
          "created_at_unix": {
            "type": "integer",
            "format": "int64"
          },
          "creator_id": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/UserId"
              }
            ]
          },
          "description": {
            "type": "string"
          },
          "id": {
            "$ref": "#/components/schemas/CollectionId"
          },
          "name": {
            "type": "string"
          },
          "public": {
            "type": "boolean"
          },
          "updated_at_unix": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
...
      "CollectionId": {
        "type": "string",
        "format": "uuid"
      },
      "CollectionPage": {
        "type": "object",
        "description": "Keyset page (ARCHITECTURE §6): pass `next_cursor` back as `cursor`.",
        "required": [
          "items"
        ],
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Collection"
            }
          },
          "next_cursor": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/CollectionId"
              }
            ]
          }
        }
      },
      "Course": {
        "type": "object",
        "required": [
          "id",
          "name",
          "description",
          "about",
          "tags",
          "public",
          "open_to_contributors",
          "created_at_unix",
          "updated_at_unix"
        ],
        "properties": {
          "about": {
            "type": "string"
          },
          "created_at_unix": {
            "type": "integer",
            "format": "int64"
          },
          "creator_id": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/UserId"
              }
            ]
          },
          "description": {
            "type": "string"
          },
          "id": {
            "$ref": "#/components/schemas/CourseId"
          },
          "name": {
            "type": "string"
          },
          "open_to_contributors": {
            "type": "boolean"
          },
          "public": {
            "type": "boolean"
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "updated_at_unix": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
      "CourseId": {
        "type": "string",
        "format": "uuid"
      },
...
      "CreateCollectionRequest": {
        "type": "object",
        "required": [
          "name"
        ],
        "properties": {
          "courses": {
            "type": [
              "array",
              "null"
            ],
            "items": {
              "$ref": "#/components/schemas/CourseId"
            },
            "description": "Course membership; every course must be readable by the caller."
          },
          "description": {
            "type": [
              "string",
              "null"
            ]
          },
          "name": {
            "type": "string"
          },
          "public": {
            "type": [
              "boolean",
              "null"
            ]
          }
        },
        "additionalProperties": false
      },
...
      "ErrorCode": {
        "type": "string",
        "description": "Stable, closed set of machine-readable error codes.",
        "enum": [
          "internal",
          "not-found",
          "forbidden",
          "unauthenticated",
          "conflict",
          "validation-failed",
          "rate-limited",
          "payload-too-large",
          "service-unavailable",
          "invalid-credentials",
          "mfa-required",
          "session-expired",
          "csrf-rejected",
          "account-disabled",
          "google-oauth-expired",
          "invalid-totp-code"
        ]
      },
      "FieldError": {
        "type": "object",
        "description": "A single field-level validation failure. `code` is a stable machine key the\nfrontend translates (e.g. `required`, `too-long`); `message` is English.",
        "required": [
          "field",
          "code",
          "message"
        ],
        "properties": {
          "code": {
            "type": "string"
          },
          "field": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        }
      },
...
      "Problem": {
        "type": "object",
        "description": "The wire shape of every error response.",
        "required": [
          "type",
          "status",
          "code",
          "title"
        ],
        "properties": {
          "code": {
            "$ref": "#/components/schemas/ErrorCode",
            "description": "Stable machine code — the frontend's i18n key."
          },
          "detail": {
            "type": [
              "string",
              "null"
            ]
          },
          "field_errors": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/FieldError"
            }
          },
          "request_id": {
            "type": [
              "string",
              "null"
            ],
            "description": "Correlation id; also present as the `x-request-id` response header."
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "title": {
            "type": "string"
          },
          "type": {
            "type": "string",
            "description": "Stable documentation URI for this error class."
          }
        }
      },
...
      "UpdateCollectionRequest": {
        "type": "object",
        "properties": {
          "courses": {
            "type": [
              "array",
              "null"
            ],
            "items": {
              "$ref": "#/components/schemas/CourseId"
            },
            "description": "Replaces the whole membership when present (legacy semantics)."
          },
          "description": {
            "type": [
              "string",
              "null"
            ]
          },
          "name": {
            "type": [
              "string",
              "null"
            ]
          },
          "public": {
            "type": [
              "boolean",
              "null"
            ]
          }
        },
        "additionalProperties": false
      },
...
      "UserId": {
        "type": "string",
        "format": "uuid"
      },
...
    }
  },
...
}